requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
web3==7.3.0
pytest==7.3.1
pytest-flask==1.2.0
